+++ packages/widget_toolkit/test/open_url_test.dart
@@ -24,7 +24,7 @@ void main() {
       void callback() => callbacksCalled++;
 
       await tester.pumpWidget(
-        OpenUrlWidget.withDependencies(
+        OpenUrlWidgetWithDependencies(
           key: openUrlKey,
           url: url,
           onTap: callback,