+++ packages/widget_toolkit_pin/CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.0]
+- Improvements to error handling
+- Errors of type `ErrorModel` thrown by `PinCodeService` are now caught by the `PinCodeKeyboard.onError` callback
+### Breaking changes:
+- Update `PinCodeService.verifyPinCode` to return a `Future<dynamic>` instead of a `bool`
+- Update `PinCodeKeyboard.onAuthenticated` signature to accept a `dynamic` parameter passed from `PinCodeService.verifyPinCode` on authentication success
+
 ## [0.0.2]
 - Upgrade major versions of dependencies: `theme_tailor`, `theme_tailor_annotation`
 ### Breaking changes: